refactor(muya): move selectAll to the Selection facade
selectAll coordinates across text and table selection (Cmd+A escalation:
cell -> table -> document). It lived in TextSelection and reached sideways into
its sibling via this._selection.table. Move it to the facade, which owns _table
directly and delegates the text-only cases (code/language clamp, same-block,
whole-document) to TextSelection.selectAllContent. TextSelection no longer
depends on its sibling table selection.
